+++ packages/spacecat-shared-gpt-client/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@adobe/spacecat-shared-gpt-client-v1.1.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-gpt-client-v1.1.0...@adobe/spacecat-shared-gpt-client-v1.1.1) (2024-02-08)
+
+
+### Bug Fixes
+
+* waiting job status ([#145](https://github.com/adobe/spacecat-shared/issues/145)) ([9fc9b6c](https://github.com/adobe/spacecat-shared/commit/9fc9b6c356f7438c52044cdc114279777f4af7ae))
+
+# [@adobe/spacecat-shared-gpt-client-v1.1.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-gpt-client-v1.0.0...@adobe/spacecat-shared-gpt-client-v1.1.0) (2024-02-08)
+
+
+### Features
+
+* switch to async firefall api ([#143](https://github.com/adobe/spacecat-shared/issues/143)) ([a431118](https://github.com/adobe/spacecat-shared/commit/a43111817c49befe66f45f16db69b7db9d469355))
+
 # @adobe/spacecat-shared-gpt-client-v1.0.0 (2024-02-06)